Renamed `Error` to `CompositionError`
