+++ src/HotChocolate/Fusion-vnext/src/Fusion.Composition/Errors/CompositionError.cs
@@ -1,0 +1,6 @@
+namespace HotChocolate.Fusion.Errors;
+
+public sealed class CompositionError(string message)
+{
+    public string Message { get; } = message;
+}
