+++ src/HotChocolate/Fusion-vnext/src/Fusion.Composition/Errors/ErrorHelper.cs
@@ -1,10 +1,10 @@
 using HotChocolate.Fusion.PreMergeValidation.Contracts;
 using static HotChocolate.Fusion.Properties.CompositionResources;
 
-namespace HotChocolate.Fusion.Results;
+namespace HotChocolate.Fusion.Errors;
 
 internal static class ErrorHelper
 {
-    public static Error PreMergeValidationRuleFailed(IPreMergeValidationRule rule)
+    public static CompositionError PreMergeValidationRuleFailed(IPreMergeValidationRule rule)
         => new(string.Format(ErrorHelper_PreMergeValidationRuleFailed, rule.GetType().Name));
 }
