--- src/HotChocolate/Fusion-vnext/src/Fusion.Composition/Results/Error.cs
@@ -1,6 +1,0 @@
-namespace HotChocolate.Fusion.Results;
-
-public sealed class Error(string message)
-{
-    public string Message { get; } = message;
-}
